Fix: change card bin type to string

+++ feature/bankcard/src/main/java/com/gorosheg/bankcard/presentation/BankCardFragment.kt
@@ -38,7 +38,7 @@ class BankCardFragment : Fragment(R.layout.fragment_bank_card) {
         bankCardMainRecycler.adapter = adapter
 
         searchIcon.setOnClickListener {
-            val cardBin = searchBin.text.toString().toInt()
+            val cardBin = searchBin.text.toString()
             viewModel.searchCard(cardBin)
         }
     }
